@@ -1,6 +1,6 @@
 {
   "name": "crowi",
-  "version": "1.5.3",
+  "version": "1.6.1",
   "description": "The simple & powerful Wiki",
   "tags": [
     "wiki",
@@ -24,19 +24,21 @@
     "url": "https://github.com/crowi/crowi.git"
   },
   "engines": {
-    "node": "4.x",
-    "npm": "3.x"
+    "node": "6.x",
+    "npm": "4.x"
   },
   "dependencies": {
     "async": "~1.5.0",
     "aws-sdk": "~2.2.26",
-    "axios": "0.13.x",
-    "babel-core": "~6.7.6",
-    "babel-loader": "~6.2.4",
-    "babel-preset-es2015": "~6.6.0",
-    "babel-preset-react": "~6.5.0",
+    "axios": "0.15.x",
+    "babel-cli": "~6.24.1",
+    "babel-core": "~6.24.1",
+    "babel-loader": "~7.0.0",
+    "babel-polyfill": "~6.23.0",
+    "babel-preset-env": "~1.4.0",
+    "babel-preset-es2015": "~6.24.1",
+    "babel-preset-react": "~6.24.1",
     "basic-auth-connect": "~1.0.0",
-    "bluebird": "~3.0.5",
     "body-parser": "~1.14.1",
     "bootstrap-sass": "~3.3.6",
     "botkit": "~0.1.1",
@@ -45,58 +47,63 @@
     "commander": "~2.9.0",
     "connect-flash": "~0.1.1",
     "connect-redis": "~2.1.0",
-    "consolidate": "~0.11.0",
+    "consolidate": "~0.14.0",
     "cookie-parser": "~1.3.4",
     "csrf": "~3.0.3",
     "debug": "~2.2.0",
     "del": "~2.2.0",
-    "diff": "~2.2.2",
-    "elasticsearch": "~11.0.1",
+    "diff": "~3.2.0",
+    "diff2html": "~2.0.12",
+    "elasticsearch": "^12.1.3",
     "emojify.js": "^1.1.0",
     "errorhandler": "~1.3.4",
-    "express": "~4.13.3",
+    "express": "~4.15.2",
     "express-form": "~0.12.0",
-    "express-session": "~1.12.0",
-    "font-awesome": "~4.5.0",
+    "express-session": "~1.14.0",
+    "font-awesome": "~4.7.0",
     "googleapis": "=12.3.0",
     "gulp": "~3.9.0",
     "gulp-concat": "~2.6.0",
     "gulp-cssmin": "~0.1.7",
     "gulp-jshint": "~1.12.0",
     "gulp-rename": "~1.2.2",
-    "gulp-sass": "~3.0.0",
+    "gulp-sass": "~3.1.0",
     "gulp-spawn-mocha": "~2.2.1",
     "gulp-uglify": "~1.4.2",
     "gulp-watch": "~4.3.5",
-    "highlight.js": "~9.0.0",
+    "highlight.js": "~9.9.0",
+    "i18next": "~4.1.0",
+    "i18next-express-middleware": "~1.0.2",
+    "i18next-node-fs-backend": "~0.1.3",
+    "i18next-sprintf-postprocessor": "~0.2.2",
     "inline-attachment": "git+https://github.com/Rovak/InlineAttachment.git#2.0.3",
     "jquery": "~2.1.4",
     "jquery.cookie": "~1.4.1",
     "jshint-stylish": "~2.1.0",
-    "kerberos": "0.0.17",
+    "kerberos": "0.0.22",
     "marked": "~0.3.5",
     "method-override": "~2.3.1",
     "mkdirp": "~0.5.1",
     "moment": "~2.13.0",
-    "mongoose": "4.2.5",
-    "mongoose-paginate": "4.2.0",
-    "morgan": "~1.5.1",
-    "multer": "~0.1.8",
-    "nodemailer": "~1.2.2",
-    "nodemailer-ses-transport": "~1.1.0",
-    "react": "~15.0.1",
-    "react-dom": "~15.0.1",
-    "redis": "~0.12.1",
+    "mongoose": "4.7.x",
+    "mongoose-paginate": "5.0.x",
+    "morgan": "~1.7.0",
+    "multer": "~1.2.1",
+    "nodemailer": "~2.7.0",
+    "nodemailer-ses-transport": "~1.5.0",
+    "react": "~15.5.0",
+    "react-bootstrap": "~0.30.10",
+    "react-dom": "~15.5.0",
+    "redis": "~2.6.5",
     "reveal.js": "~3.2.0",
     "socket.io": "~1.3.0",
     "socket.io-client": "~1.3.0",
     "sprintf": "~0.1.5",
     "swig": "~1.4.0",
-    "time": "~0.11.0",
     "vinyl-source-stream": "~1.1.0",
-    "webpack": "~1.13.0",
-    "webpack-manifest-plugin": "~1.0.1",
-    "webpack-stream": "~3.1.0"
+    "webpack": "~2.2.0",
+    "webpack-manifest-plugin": "~1.1.0",
+    "webpack-stream": "~3.2.0"
   },
   "devDependencies": {
     "chai": "~1.10.0",
