@@ -4,7 +4,6 @@ module.exports = function(crowi, app) {
   var debug = require('debug')('crowi:crowi:express-init')
     , express        = require('express')
     , bodyParser     = require('body-parser')
-    , multer         = require('multer')
     , cookieParser   = require('cookie-parser')
     , methodOverride = require('method-override')
     , session        = require('express-session')
@@ -12,9 +11,36 @@ module.exports = function(crowi, app) {
     , flash          = require('connect-flash')
     , cons           = require('consolidate')
     , swig           = require('swig')
+    , i18next        = require('i18next')
+    , i18nFsBackend  = require('i18next-node-fs-backend')
+    , i18nSprintf    = require('i18next-sprintf-postprocessor')
+    , i18nMiddleware = require('i18next-express-middleware')
+    , i18nUserSettingDetector  = require('../util/i18nUserSettingDetector')
     , env            = crowi.node_env
     , middleware     = require('../util/middlewares')
+
+    , User = crowi.model('User')
     ;
+
+  var lngDetector = new i18nMiddleware.LanguageDetector();
+  lngDetector.addDetector(i18nUserSettingDetector);
+
+  i18next
+    .use(lngDetector)
+    .use(i18nFsBackend)
+    .use(i18nSprintf)
+    .init({
+      debug: (crowi.node_env === 'development'),
+      fallbackLng: [User.LANG_EN_US],
+      whitelist: Object.keys(User.getLanguageLabels()).map((k) => User[k]),
+      backend: {
+        loadPath: crowi.localeDir + '{{lng}}/translation.json'
+      },
+      detection: {
+        order: ['userSettingDetector', 'header', 'navigator'],
+      },
+      overloadTranslationOptionHandler: i18nSprintf.overloadTranslationOptionHandler
+    });
 
   app.use(function(req, res, next) {
     var now = new Date()
@@ -42,8 +68,10 @@ module.exports = function(crowi, app) {
     res.locals.consts   = {
         pageGrants: Page.getGrantLabels(),
         userStatus: User.getUserStatusLabels(),
+        language:   User.getLanguageLabels(),
         registrationMode: Config.getRegistrationModeLabels(),
     };
+    res.locals.local_config = Config.getLocalconfig(config); // config for browser context
 
     next();
   });
@@ -71,7 +99,6 @@ module.exports = function(crowi, app) {
   app.use(bodyParser.urlencoded({ extended: true, limit: '50mb' }));
   app.use(bodyParser.json({limit: '50mb'}));
   app.use(cookieParser());
-  app.use(multer());
   app.use(session(crowi.sessionConfig));
   app.use(flash());
 
@@ -80,4 +107,5 @@ module.exports = function(crowi, app) {
 
   app.use(middleware.loginChecker(crowi, app));
 
+  app.use(i18nMiddleware.handle(i18next));
 };
