@@ -6,7 +6,6 @@ var debug = require('debug')('crowi:crowi')
   , path = require('path')
   , fs = require('fs')
   , sep = path.sep
-  , Promise = require('bluebird')
 
   , mongoose    = require('mongoose')
 
@@ -24,6 +23,7 @@ function Crowi (rootdir, env)
   this.publicDir = path.join(this.rootDir, 'public') + sep;
   this.libDir    = path.join(this.rootDir, 'lib') + sep;
   this.eventsDir = path.join(this.libDir, 'events') + sep;
+  this.localeDir = path.join(this.rootDir, 'locales') + sep;
   this.resourceDir = path.join(this.rootDir, 'resource') + sep;
   this.viewsDir  = path.join(this.libDir, 'views') + sep;
   this.mailDir   = path.join(this.viewsDir, 'mail') + sep;
@@ -55,16 +55,13 @@ function Crowi (rootdir, env)
   };
 
   if (this.node_env == 'development') {
-    Promise.longStackTraces();
   }
-
-  //time.tzset('Asia/Tokyo');
 };
 
 Crowi.prototype.init = function() {
   var self = this;
 
-  return new Promise.resolve()
+  return Promise.resolve()
   .then(function() {
     // setup database server and load all modesl
     return self.setupDatabase();
@@ -94,10 +91,20 @@ Crowi.prototype.init = function() {
     return self.setupSlack();
   }).then(function() {
     return self.setupCsrf();
-  }).then(function() {
-    return self.buildServer();
   });
 }
+
+Crowi.prototype.isPageId = function(pageId) {
+  if (!pageId) {
+    return false;
+  }
+
+  if (typeof pageId === 'string' && pageId.match(/^[\da-f]{24}$/)) {
+    return true;
+  }
+
+  return false;
+};
 
 Crowi.prototype.setConfig = function(config) {
   this.config = config;
@@ -143,6 +150,8 @@ Crowi.prototype.event = function(name, event) {
 
 Crowi.prototype.setupDatabase = function() {
   // mongoUri = mongodb://user:password@host/dbname
+  mongoose.Promise = global.Promise;
+
   var mongoUri = this.env.MONGOLAB_URI || // for B.C.
     this.env.MONGODB_URI || // MONGOLAB changes their env name
     this.env.MONGOHQ_URL ||
@@ -328,7 +337,7 @@ Crowi.prototype.buildServer = function() {
     });
   }
 
-  return new Promise.resolve(app);
+  return Promise.resolve(app);
 };
 
 Crowi.prototype.exitOnError = function(err) {
