@@ -5,6 +5,7 @@ var form = require('express-form')
 
 module.exports = form(
   field('settingForm[app:title]').required(),
-  field('settingForm[app:confidential]')
+  field('settingForm[app:confidential]'),
+  field('settingForm[app:fileUpload]').isInt()
 );
 
