@@ -6,6 +6,6 @@ var form = require('express-form')
 module.exports = form(
   field('invitedForm.username').required().is(/^[\da-zA-Z\-_]+$/),
   field('invitedForm.name').required(),
-  field('invitedForm.password').required().is(/^[\da-zA-Z@#$%-_&\+\*\?]{6,64}$/)
+  field('invitedForm.password').required().is(/^[\x20-\x7F]{6,}$/)
 );
 
