@@ -5,5 +5,5 @@ var form = require('express-form')
 
 module.exports = form(
   field('loginForm.email').required(),
-  field('loginForm.password').required().is(/^[\x20-\x7F]{6,40}$/)
+  field('loginForm.password').required().is(/^[\x20-\x7F]{6,}$/)
 );
