@@ -5,5 +5,6 @@ var form = require('express-form')
 
 module.exports = form(
   field('userForm.name').trim().required(),
-  field('userForm.email').trim().isEmail().required()
+  field('userForm.email').trim().isEmail().required(),
+  field('userForm.lang').required()
 );
