@@ -2,7 +2,6 @@ module.exports = function(crowi) {
   var debug = require('debug')('crowi:models:attachment')
     , mongoose = require('mongoose')
     , ObjectId = mongoose.Schema.Types.ObjectId
-    , Promise = require('bluebird')
     , fileUploader = require('../util/fileUploader')(crowi)
   ;
 
@@ -25,11 +24,11 @@ module.exports = function(crowi) {
   }, {
     toJSON: {
       virtuals: true
-    }
+    },
   });
 
   attachmentSchema.virtual('fileUrl').get(function() {
-    return fileUploader.generateUrl(this.filePath);
+    return `/files/${this._id}`;
   });
 
   attachmentSchema.statics.findById = function(id) {
@@ -101,7 +100,7 @@ module.exports = function(crowi) {
   };
 
   attachmentSchema.statics.createAttachmentFilePath = function (pageId, fileName, fileType) {
-    var ext = '.' + fileName.match(/(.*)(?:\.([^.]+$))/)[2];
+    var ext = '.' + fileName.match(/(.*)(?:\.([^.]+$))/)[2] || '';
 
     return 'attachment/' + pageId + '/' + generateFileHash(fileName) + ext;
   };
@@ -109,37 +108,52 @@ module.exports = function(crowi) {
   attachmentSchema.statics.removeAttachmentsByPageId = function(pageId) {
     var Attachment = this;
 
-    // todo
-    return Promise.resolve();
-    //return new Promise(function(resolve, reject) {
-    //  // target find
-    //  Attachment.getListByPageId(pageId)
-    //  .then(function(attachments) {
-    //  }).then(function(done) {
-    //  }).catch(function(err) {
-    //  });
-    //});
+    return new Promise((resolve, reject) => {
+      Attachment.getListByPageId(pageId)
+      .then((attachments) => {
+        for (attachment of attachments) {
+          Attachment.removeAttachment(attachment).then((res) => {
+            // do nothing
+          }).catch((err) => {
+            debug('Attachment remove error', err);
+          });
+        }
+
+        resolve(attachments);
+      }).catch((err) => {
+        reject(err);
+      });
+    });
 
   };
 
-  attachmentSchema.statics.createCacheFileName = function(attachment) {
-    return crowi.cacheDir + 'attachment-' + attachment._id;
+  attachmentSchema.statics.findDeliveryFile = function(attachment, forceUpdate) {
+    var Attachment = this;
+
+    // TODO
+    var forceUpdate = forceUpdate || false;
+
+    return fileUploader.findDeliveryFile(attachment._id, attachment.filePath);
   };
 
-  attachmentSchema.statics.findDeliveryFile = function(attachment) {
-    // find local
-    var fs = require('fs');
-    var deliveryFile = {
-      filename: '',
-      options: {
-        headers: {
-          'Content-Type': attachment.fileFormat,
-        },
-      },
-    };
-    var cacheFileName = this.createCacheFileName(attachment);
-    // とちゅう
-    return deliveryFile;
+  attachmentSchema.statics.removeAttachment = function(attachment) {
+    const Attachment = this;
+    const filePath = attachment.filePath;
+
+    return new Promise((resolve, reject) => {
+      Attachment.remove({_id: attachment._id}, (err, data) => {
+        if (err) {
+          return reject(err);
+        }
+
+        fileUploader.deleteFile(attachment._id, filePath)
+        .then(data => {
+          resolve(data); // this may null
+        }).catch(err => {
+          reject(err);
+        });
+      });
+    });
   };
 
   return mongoose.model('Attachment', attachmentSchema);
