@@ -329,10 +329,10 @@ module.exports = function(crowi) {
 
   pageSchema.statics.getGrantLabels = function() {
     var grantLabels = {};
-    grantLabels[GRANT_PUBLIC]     = '公開';
-    grantLabels[GRANT_RESTRICTED] = 'リンクを知っている人のみ';
-    //grantLabels[GRANT_SPECIFIED]  = '特定ユーザーのみ';
-    grantLabels[GRANT_OWNER]      = '自分のみ';
+    grantLabels[GRANT_PUBLIC]     = 'Public'; // 公開
+    grantLabels[GRANT_RESTRICTED] = 'Anyone with the link'; // リンクを知っている人のみ
+    //grantLabels[GRANT_SPECIFIED]  = 'Specified users only'; // 特定ユーザーのみ
+    grantLabels[GRANT_OWNER]      = 'Just me'; // 自分のみ
 
     return grantLabels;
   };
@@ -341,6 +341,8 @@ module.exports = function(crowi) {
     if (!path.match(/^\//)) {
       path = '/' + path;
     }
+
+    path = path.replace(/\/\s+?/g, '/').replace(/\s+\//g, '/');
 
     return path;
   };
@@ -383,6 +385,8 @@ module.exports = function(crowi) {
       /^\/_r\/.*/,
       /^\/user\/[^\/]+\/(bookmarks|comments|activities|pages|recent-create|recent-edit)/, // reserved
       /^\/?https?:\/\/.+$/, // avoid miss in renaming
+      /\/{2,}/,             // avoid miss in renaming
+      /\s+\/\s+/,           // avoid miss in renaming
       /.+\/edit$/,
       /.+\.md$/,
       /^\/(installer|register|login|logout|admin|me|files|trash|paste|comments)(\/.*|$)/,
@@ -569,6 +573,20 @@ module.exports = function(crowi) {
     });
   };
 
+  pageSchema.statics.findPageByRedirectTo = function(path) {
+    var Page = this;
+
+    return new Promise(function(resolve, reject) {
+      Page.findOne({redirectTo: path}, function(err, pageData) {
+        if (err || pageData === null) {
+          return reject(err);
+        }
+
+        return resolve(pageData);
+      });
+    });
+  };
+
   pageSchema.statics.findListByCreator = function(user, option, currentUser) {
     var Page = this;
     var User = crowi.model('User');
@@ -613,7 +631,7 @@ module.exports = function(crowi) {
         {path: 'revision', model: 'Revision'},
       ])
       .sort({updatedAt: -1})
-      .stream();
+      .cursor();
   };
 
   /**
@@ -931,8 +949,10 @@ module.exports = function(crowi) {
       }).then(function(done) {
         return Page.removePageById(pageId);
       }).then(function(done) {
+        return Page.removeRedirectOriginPageByPath(pageData.path);
+      }).then(function(done) {
         pageEvent.emit('delete', pageData, user); // update as renamed page
-        resolve();
+        resolve(pageData);
       }).catch(reject);
     });
   };
@@ -950,6 +970,43 @@ module.exports = function(crowi) {
         resolve(done);
       });
     });
+  };
+
+  pageSchema.statics.removePageByPath = function(pagePath) {
+    var Page = this;
+
+    return Page.findPageByPath(redirectPath)
+      .then(function(pageData) {
+        return Page.removePageById(pageData.id);
+      });
+  };
+
+  /**
+   * remove the page that is redirecting to specified `pagePath` recursively
+   *  ex: when
+   *    '/page1' redirects to '/page2' and
+   *    '/page2' redirects to '/page3'
+   *    and given '/page3',
+   *    '/page1' and '/page2' will be removed
+   *
+   * @param {string} pagePath
+   */
+  pageSchema.statics.removeRedirectOriginPageByPath = function(pagePath) {
+    var Page = this;
+
+    return Page.findPageByRedirectTo(pagePath)
+      .then((redirectOriginPageData) => {
+        // remove
+        return Page.removePageById(redirectOriginPageData.id)
+          // remove recursive
+          .then(() => {
+            return Page.removeRedirectOriginPageByPath(redirectOriginPageData.path)
+          });
+      })
+      .catch((err) => {
+        // do nothing if origin page doesn't exist
+        return Promise.resolve();
+      })
   };
 
   pageSchema.statics.rename = function(pageData, newPagePath, user, options) {
