@@ -11,7 +11,12 @@ module.exports = function(crowi) {
     , STATUS_SUSPENDED  = 3
     , STATUS_DELETED    = 4
     , STATUS_INVITED    = 5
-    , USER_PUBLIC_FIELDS = '_id image googleId name username email introduction status createdAt admin' // TODO: どこか別の場所へ...
+    , USER_PUBLIC_FIELDS = '_id image googleId name username email introduction status lang createdAt admin' // TODO: どこか別の場所へ...
+
+    , LANG_EN    = 'en'
+    , LANG_EN_US = 'en-US'
+    , LANG_EN_GB = 'en-GB'
+    , LANG_JA    = 'ja'
 
     , PAGE_ITEMS        = 50
 
@@ -29,6 +34,11 @@ module.exports = function(crowi) {
     introduction: { type: String },
     password: String,
     apiToken: String,
+    lang: {
+      type: String,
+      enum: Object.keys(getLanguageLabels()).map((k) => eval(k)),
+      default: LANG_EN_US
+    },
     status: { type: Number, required: true, default: STATUS_ACTIVE, index: true  },
     createdAt: { type: Date, default: Date.now },
     admin: { type: Boolean, default: 0, index: true  }
@@ -84,6 +94,16 @@ module.exports = function(crowi) {
     return hasher.digest('base64');
   }
 
+  function getLanguageLabels() {
+    var lang = {};
+    lang.LANG_EN    = LANG_EN;
+    lang.LANG_EN_US = LANG_EN_US;
+    lang.LANG_EN_GB = LANG_EN_GB;
+    lang.LANG_JA    = LANG_JA;
+
+    return lang;
+  }
+
   userSchema.methods.isPasswordSet = function() {
     if (this.password) {
       return true;
@@ -107,9 +127,11 @@ module.exports = function(crowi) {
     return false;
   };
 
-  userSchema.methods.update = function(name, email, callback) {
+  userSchema.methods.update = function(name, email, lang, callback) {
     this.name = name;
     this.email = email;
+    this.lang = lang;
+
     this.save(function(err, userData) {
       return callback(err, userData);
     });
@@ -232,6 +254,7 @@ module.exports = function(crowi) {
     });
   };
 
+  userSchema.statics.getLanguageLabels = getLanguageLabels;
   userSchema.statics.getUserStatusLabels = function() {
     var userStatus = {};
     userStatus[STATUS_REGISTERED] = '承認待ち';
@@ -292,13 +315,18 @@ module.exports = function(crowi) {
     var User = this;
     var option = option || {}
       , sort = option.sort || {createdAt: -1}
-      , status = option.status || STATUS_ACTIVE
+      , status = option.status || [STATUS_ACTIVE, STATUS_SUSPENDED]
       , fields = option.fields || USER_PUBLIC_FIELDS
       ;
 
+    if (!Array.isArray(status)) {
+      status = [status];
+    }
+
     return new Promise(function(resolve, reject) {
       User
-        .find({status: status })
+        .find()
+        .or(status.map(s => { return {status: s}; }))
         .select(fields)
         .sort(sort)
         .exec(function (err, userData) {
@@ -347,13 +375,13 @@ module.exports = function(crowi) {
   userSchema.statics.findUsersWithPagination = function(options, callback) {
     var sort = options.sort || {status: 1, username: 1, createdAt: 1};
 
-    this.paginate({}, { page: options.page || 1, limit: PAGE_ITEMS }, function(err, paginatedResults, pageCount, itemCount) {
+    this.paginate({}, { page: options.page || 1, limit: options.limit || PAGE_ITEMS }, function(err, result) {
       if (err) {
         debug('Error on pagination:', err);
         return callback(err, null);
       }
 
-      return callback(err, paginatedResults, pageCount, itemCount);
+      return callback(err, result);
     }, { sortBy : sort });
   };
 
@@ -612,7 +640,7 @@ module.exports = function(crowi) {
     );
   };
 
-  userSchema.statics.createUserByEmailAndPassword = function(name, username, email, password, callback) {
+  userSchema.statics.createUserByEmailAndPassword = function(name, username, email, password, lang, callback) {
     var User = this
       , newUser = new User();
 
@@ -620,6 +648,7 @@ module.exports = function(crowi) {
     newUser.username = username;
     newUser.email = email;
     newUser.setPassword(password);
+    newUser.lang = lang;
     newUser.createdAt = Date.now();
     newUser.status = decideUserStatusOnRegistration();
 
@@ -647,13 +676,18 @@ module.exports = function(crowi) {
   };
 
 
-  userSchema.statics.STATUS_REGISTERED = STATUS_REGISTERED;
-  userSchema.statics.STATUS_ACTIVE = STATUS_ACTIVE;
-  userSchema.statics.STATUS_SUSPENDED = STATUS_SUSPENDED;
-  userSchema.statics.STATUS_DELETED = STATUS_DELETED;
-  userSchema.statics.STATUS_INVITED = STATUS_INVITED;
+  userSchema.statics.STATUS_REGISTERED  = STATUS_REGISTERED;
+  userSchema.statics.STATUS_ACTIVE      = STATUS_ACTIVE;
+  userSchema.statics.STATUS_SUSPENDED   = STATUS_SUSPENDED;
+  userSchema.statics.STATUS_DELETED     = STATUS_DELETED;
+  userSchema.statics.STATUS_INVITED     = STATUS_INVITED;
   userSchema.statics.USER_PUBLIC_FIELDS = USER_PUBLIC_FIELDS;
-  userSchema.statics.PAGE_ITEMS = PAGE_ITEMS;
+  userSchema.statics.PAGE_ITEMS         = PAGE_ITEMS;
+
+  userSchema.statics.LANG_EN            = LANG_EN;
+  userSchema.statics.LANG_EN_US         = LANG_EN_US;
+  userSchema.statics.LANG_EN_GB         = LANG_EN_US;
+  userSchema.statics.LANG_JA            = LANG_JA;
 
   return mongoose.model('User', userSchema);
 };
