@@ -14,6 +14,7 @@ module.exports = function(crowi, app) {
 
   actions.createAdmin = function(req, res) {
     var registerForm = req.body.registerForm || {};
+    var language = req.language || 'en';
 
     if (req.form.isValid) {
       var name = registerForm.name;
@@ -21,7 +22,7 @@ module.exports = function(crowi, app) {
       var email = registerForm.email;
       var password = registerForm.password;
 
-      User.createUserByEmailAndPassword(name, username, email, password, function(err, userData) {
+      User.createUserByEmailAndPassword(name, username, email, password, language, function(err, userData) {
         if (err) {
           req.form.errors.push('管理ユーザーの作成に失敗しました。' + err.message);
           // TODO
@@ -38,7 +39,7 @@ module.exports = function(crowi, app) {
             // login処理
             req.user = req.session.user = userData;
             req.flash('successMessage', 'Crowi のインストールが完了しました！はじめに、このページでこの Wiki の各種設定を確認してください。');
-            return res.redirect('admin/app');
+            return res.redirect('/admin/app');
           });
         });
       });
