@@ -39,7 +39,7 @@ module.exports = function(crowi, app) {
   };
 
   var loginFailure = function(req, res) {
-    req.flash('warningMessage', 'ログインに失敗しました');
+    req.flash('warningMessage', 'Sign in failure.');
     return res.redirect('/login');
   };
 
@@ -59,9 +59,9 @@ module.exports = function(crowi, app) {
       ;
 
     if (reason === 'suspended') {
-      reasonMessage = 'このアカウントは停止されています。';
+      reasonMessage = 'This account is suspended.';
     } else if (reason === 'registered') {
-      reasonMessage = '管理者の承認をお待ちください。';
+      reasonMessage = 'Wait for approved by administrators.';
     } else {
     }
 
@@ -129,6 +129,7 @@ module.exports = function(crowi, app) {
 
   actions.register = function(req, res) {
     var googleAuth = require('../util/googleAuth')(config);
+    var lang= req.lang || User.LANG_EN_US;
 
     // ログイン済みならさようなら
     if (req.user) {
@@ -155,16 +156,16 @@ module.exports = function(crowi, app) {
         var isError = false;
         if (!User.isEmailValid(email)) {
           isError = true;
-          req.flash('registerWarningMessage', 'このメールアドレスは登録できません。(ホワイトリストなどを確認してください)');
+          req.flash('registerWarningMessage', 'This email address could not be used. (Make sure the allowed email address)');
         }
         if (!isRegisterable) {
           if (!errOn.username) {
             isError = true;
-            req.flash('registerWarningMessage', 'このユーザーIDは利用できません。');
+            req.flash('registerWarningMessage', 'This User ID is not available.');
           }
           if (!errOn.email) {
             isError = true;
-            req.flash('registerWarningMessage', 'このメールアドレスは登録済みです。');
+            req.flash('registerWarningMessage', 'This email address is already registered.');
           }
 
         }
@@ -173,9 +174,9 @@ module.exports = function(crowi, app) {
           return res.redirect('/register');
         }
 
-        User.createUserByEmailAndPassword(name, username, email, password, function(err, userData) {
+        User.createUserByEmailAndPassword(name, username, email, password, lang, function(err, userData) {
           if (err) {
-            req.flash('registerWarningMessage', 'ユーザー登録に失敗しました。');
+            req.flash('registerWarningMessage', 'Failed to register.');
             return res.redirect('/register');
           } else {
 
@@ -272,7 +273,7 @@ module.exports = function(crowi, app) {
           req.session.googleAuthCode = null;
 
           if (err) {
-            req.flash('registerWarningMessage', 'Googleコネクト中にエラーが発生しました。');
+            req.flash('registerWarningMessage', 'Error on connectiong Google');
             return res.redirect('/login?register=1'); // TODO Handling
           }
 
