@@ -22,11 +22,40 @@ module.exports = function(crowi, app) {
     Revision
       .findRevision(revisionId)
       .then(function(revisionData) {
-        return res.json(ApiResponse.success(revisionData));
+        var result = {
+          revision: revisionData,
+        }
+        return res.json(ApiResponse.success(result));
       })
       .catch(function(err) {
+        debug('Error revisios.get', err);
         return res.json(ApiResponse.error(err));
       });
+  };
+
+  /**
+   * @api {get} /revisions.ids Get revision id list of the page
+   * @apiName ids
+   * @apiGroup Revision
+   *
+   * @apiParam {String} page_id      Page Id.
+   */
+  actions.api.ids = function(req, res) {
+    var pageId = req.query.page_id || null;
+
+    if (pageId && crowi.isPageId(pageId)) {
+      Page.findPageByIdAndGrantedUser(pageId, req.user)
+      .then(function(pageData) {
+        debug('Page found', pageData._id, pageData.path);
+        return Revision.findRevisionIdList(pageData.path);
+      }).then(function(revisions) {
+        return res.json(ApiResponse.success({revisions}));
+      }).catch(function(err) {
+        return res.json(ApiResponse.error(err));
+      });
+    } else {
+      return res.json(ApiResponse.error('Parameter error.'));
+    }
   };
 
   /**
