@@ -19,6 +19,7 @@ function SearchClient(crowi, esUri) {
   this.client = new elasticsearch.Client({
     host: this.host,
     requestTimeout: 5000,
+    //log: 'debug',
   });
 
   this.registerUpdateEvent();
@@ -208,11 +209,13 @@ SearchClient.prototype.addAllPages = function()
   var self = this;
   var offset = 0;
   var Page = this.crowi.model('Page');
-  var stream = Page.getStreamOfFindAll();
+  var cursor = Page.getStreamOfFindAll();
   var body = [];
 
+  var counter = 0;
+
   return new Promise(function(resolve, reject) {
-    stream.on('data', function (doc) {
+    cursor.on('data', function (doc) {
       if (!doc.creator || !doc.revision || !self.shouldIndexed(doc)) {
         debug('Skipped', doc.path);
         return ;
@@ -221,7 +224,7 @@ SearchClient.prototype.addAllPages = function()
       self.prepareBodyForCreate(body, doc);
     }).on('error', function (err) {
       // TODO: handle err
-      debug('Error stream:', err);
+      debug('Error cursor:', err);
     }).on('close', function () {
       // all done
 
@@ -276,7 +279,7 @@ SearchClient.prototype.search = function(query)
 SearchClient.prototype.createSearchQuerySortedByUpdatedAt = function(option)
 {
   // getting path by default is almost for debug
-  var fields = ['path', '_id'];
+  var fields = ['path'];
   if (option) {
     fields = option.fields || fields;
   }
@@ -286,9 +289,9 @@ SearchClient.prototype.createSearchQuerySortedByUpdatedAt = function(option)
     index: this.index_name,
     type: 'pages',
     body: {
-      fields: fields,
       sort: [{ updated_at: { order: 'desc'}}],
       query: {}, // query
+      _source: fields,
     }
   };
   this.appendResultSize(query);
@@ -298,7 +301,7 @@ SearchClient.prototype.createSearchQuerySortedByUpdatedAt = function(option)
 
 SearchClient.prototype.createSearchQuerySortedByScore = function(option)
 {
-  var fields = ['path', '_id'];
+  var fields = ['path'];
   if (option) {
     fields = option.fields || fields;
   }
@@ -308,9 +311,9 @@ SearchClient.prototype.createSearchQuerySortedByScore = function(option)
     index: this.index_name,
     type: 'pages',
     body: {
-      fields: fields,
       sort: [ {_score: { order: 'desc'} }],
       query: {}, // query
+      _source: fields,
     }
   };
   this.appendResultSize(query);
@@ -330,21 +333,88 @@ SearchClient.prototype.appendCriteriaForKeywordContains = function(query, keywor
   if (!query.body.query.bool) {
     query.body.query.bool = {};
   }
-
   if (!query.body.query.bool.must || !Array.isArray(query.body.query.must)) {
     query.body.query.bool.must = [];
   }
+  if (!query.body.query.bool.must_not || !Array.isArray(query.body.query.must_not)) {
+    query.body.query.bool.must_not = [];
+  }
 
-  query.body.query.bool.must.push({
-    multi_match: {
-      query: keyword,
-      fields: [
-        "path.ja^2", // ためしに。
-        "body.ja"
-      ],
-      operator: "and"
+  var appendMultiMatchQuery = function(query, type, keywords) {
+    var target;
+    var operator = 'and';
+    switch (type) {
+      case 'not_match':
+        target = query.body.query.bool.must_not;
+        operator = 'or';
+        break;
+      case 'match':
+      default:
+        target = query.body.query.bool.must;
     }
-  });
+
+    target.push({
+      multi_match: {
+        query: keywords.join(' '),
+        // TODO: By user's i18n setting, change boost or search target fields
+        fields: [
+          "path_ja^2",
+          "body_ja",
+          // "path_en",
+          // "body_en",
+        ],
+        operator: operator,
+      }
+    });
+
+    return query;
+  };
+
+  var parsedKeywords = this.getParsedKeywords(keyword);
+
+  if (parsedKeywords.match.length > 0) {
+    query = appendMultiMatchQuery(query, 'match', parsedKeywords.match);
+  }
+
+  if (parsedKeywords.not_match.length > 0) {
+    query = appendMultiMatchQuery(query, 'not_match', parsedKeywords.not_match);
+  }
+
+  if (parsedKeywords.phrase.length > 0) {
+    var phraseQueries = [];
+    parsedKeywords.phrase.forEach(function(phrase) {
+      phraseQueries.push({
+        multi_match: {
+          query: phrase, // each phrase is quoteted words
+          type: 'phrase',
+          fields: [ // Not use "*.ja" fields here, because we want to analyze (parse) search words
+            "path_raw^2",
+            "body_raw",
+          ],
+        }
+      });
+    });
+
+    query.body.query.bool.must.push(phraseQueries);
+  }
+
+  if (parsedKeywords.not_phrase.length > 0) {
+    var notPhraseQueries = [];
+    parsedKeywords.not_phrase.forEach(function(phrase) {
+      notPhraseQueries.push({
+        multi_match: {
+          query: phrase, // each phrase is quoteted words
+          type: 'phrase',
+          fields: [ // Not use "*.ja" fields here, because we want to analyze (parse) search words
+            "path_raw^2",
+            "body_raw",
+          ],
+        }
+      });
+    });
+
+    query.body.query.bool.must_not.push(notPhraseQueries);
+  }
 };
 
 SearchClient.prototype.appendCriteriaForPathFilter = function(query, path)
@@ -363,7 +433,7 @@ SearchClient.prototype.appendCriteriaForPathFilter = function(query, path)
   }
   query.body.query.bool.filter.push({
     wildcard: {
-      "path.raw": path + "/*"
+      "path": path + "/*"
     }
   });
 };
@@ -395,6 +465,54 @@ SearchClient.prototype.searchKeywordUnderPath = function(keyword, path, option)
 
   return this.search(query);
 };
+
+SearchClient.prototype.getParsedKeywords = function(keyword)
+{
+  var matchWords = [];
+  var notMatchWords = [];
+  var phraseWords = [];
+  var notPhraseWords = [];
+
+  keyword.trim();
+  keyword = keyword.replace(/\s+/g, ' ');
+
+  // First: Parse phrase keywords
+  var phraseRegExp = new RegExp(/(-?"[^"]+")/g);
+  var phrases = keyword.match(phraseRegExp);
+
+  if (phrases !== null) {
+    keyword = keyword.replace(phraseRegExp, '');
+
+    phrases.forEach(function(phrase) {
+      phrase.trim();
+      if (phrase.match(/^\-/)) {
+        notPhraseWords.push(phrase.replace(/^\-/, ''));
+      } else {
+        phraseWords.push(phrase);
+      }
+    });
+  }
+
+  // Second: Parse other keywords (include minus keywords)
+  keyword.split(' ').forEach(function(word) {
+    if (word === '') {
+      return;
+    }
+
+    if (word.match(/^\-(.+)$/)) {
+      notMatchWords.push((RegExp.$1));
+    } else {
+      matchWords.push(word);
+    }
+  });
+
+  return {
+    match: matchWords,
+    not_match: notMatchWords,
+    phrase: phraseWords,
+    not_phrase: notPhraseWords,
+  };
+}
 
 SearchClient.prototype.syncPageCreated = function(page, user)
 {
