@@ -5,36 +5,53 @@ import Crowi from './util/Crowi';
 import CrowiRenderer from './util/CrowiRenderer';
 
 import HeaderSearchBox  from './components/HeaderSearchBox';
-import SearchPage  from './components/SearchPage';
-import PageListSearch  from './components/PageListSearch';
-//import PageComment  from './components/PageComment';
-import SeenUserList from './components/SeenUserList';
+import SearchPage       from './components/SearchPage';
+import PageListSearch   from './components/PageListSearch';
+import PageHistory      from './components/PageHistory';
+import PageAttachment   from './components/PageAttachment';
+import SeenUserList     from './components/SeenUserList';
+import BookmarkButton   from './components/BookmarkButton';
 import Backlink from './components/Backlink';
+//import PageComment  from './components/PageComment';
 
 if (!window) {
   window = {};
 }
+
+const mainContent = document.querySelector('#content-main');
+let pageId = null;
+let pageContent = null;
+if (mainContent !== null) {
+  pageId = mainContent.attributes['data-page-id'].value;
+  const rawText = document.getElementById('raw-text-original');
+  if (rawText) {
+    pageContent = rawText.innerHTML;
+  }
+}
+
 // FIXME
-const crowi = new Crowi({me: $('#content-main').data('current-username')}, window);
+const crowi = new Crowi({
+  me: $('#content-main').data('current-username'),
+  csrfToken: $('#content-main').data('csrftoken'),
+}, window);
 window.crowi = crowi;
+crowi.setConfig(JSON.parse(document.getElementById('crowi-context-hydrate').textContent || '{}'));
 crowi.fetchUsers();
 
 const crowiRenderer = new CrowiRenderer();
 window.crowiRenderer = crowiRenderer;
 
-const contentMain = document.querySelector('#content-main');
-let pageId = null;
-if (contentMain !== null) {
-  pageId = contentMain.dataset.pageId;
-}
-
 const componentMappings = {
-  'search-top': <HeaderSearchBox />,
-  'search-page': <SearchPage />,
-  'page-list-search': <PageListSearch />,
+  'search-top': <HeaderSearchBox crowi={crowi} />,
+  'search-page': <SearchPage crowi={crowi} />,
+  'page-list-search': <PageListSearch crowi={crowi} />,
+  'page-attachment': <PageAttachment pageId={pageId} pageContent={pageContent} crowi={crowi} />,
+
+  //'revision-history': <PageHistory pageId={pageId} />,
   //'page-comment': <PageComment />,
   'seen-user-list': <SeenUserList />,
   'backlink-list': <Backlink pageId={pageId} crowi={crowi} />,
+  'bookmark-button': <BookmarkButton pageId={pageId} crowi={crowi} />,
 };
 
 Object.keys(componentMappings).forEach((key) => {
@@ -42,4 +59,9 @@ Object.keys(componentMappings).forEach((key) => {
   if (elem) {
     ReactDOM.render(componentMappings[key], elem);
   }
+});
+
+// うわーもうー
+$('a[data-toggle="tab"][href="#revision-history"]').on('show.bs.tab', function() {
+  ReactDOM.render(<PageHistory pageId={pageId} crowi={crowi} />, document.getElementById('revision-history'));
 });
