@@ -2,7 +2,6 @@ var chai = require('chai')
   , expect = chai.expect
   , sinon = require('sinon')
   , sinonChai = require('sinon-chai')
-  , Promise = require('bluebird')
   , utils = require('../utils.js')
   ;
 chai.use(sinonChai);
@@ -146,6 +145,7 @@ describe('Page', function () {
       expect(Page.isCreatableName('http://demo.crowi.wiki/user/sotarok/hoge')).to.be.false;
       expect(Page.isCreatableName('https://demo.crowi.wiki/user/sotarok/hoge')).to.be.false;
 
+      expect(Page.isCreatableName('/ the / path / with / space')).to.be.false;
 
       var forbidden = ['installer', 'register', 'login', 'logout', 'admin', 'files', 'trash', 'paste', 'comments'];
       for (var i = 0; i < forbidden.length ; i++) {
@@ -262,6 +262,20 @@ describe('Page', function () {
         });
       });
 
+    });
+  });
+
+  describe('Normalize path', function () {
+    context('Normalize', function() {
+      it('should start with slash', function(done) {
+        expect(Page.normalizePath('hoge/fuga')).to.equal('/hoge/fuga');
+        done();
+      });
+
+      it('should trim spaces of slash', function(done) {
+        expect(Page.normalizePath('/ hoge / fuga')).to.equal('/hoge/fuga');
+        done();
+      });
     });
   });
 
